Thoroughly reviewed m-of-n code [RIPD-480]
This commit represents the portion of the m-of-n code that
Vinnie, David, and Nik gave a thorough review.  Identifying
it as one piece is intended as a guidepost for where future
reviewers should put their primary attention.

This commit does not address the following review comments:

 o sfSigningAccount must not be legal in a serialized transaction.
   Use only JSON to return the SigningAccount JSON object in RPC.

 o SigningAccount lacks an STObject template.  Identify where to
   keep such a template, put it there, and use it.

 o Validate a SigningAccount and SignerEntry objects using
   STObject::setType ().

 o Move extensive javadoc from SetSignerList.cpp into a
   transactors/README.md.

 o Fees and reserves are missing.  In particular, SetSignerList
   doesn't check if the account has enough XRP to bump the reserve
   count.  Other fee/reserve related stuff is missing as well.

There was also discussion of problems handling TER codes.  The
conclusion was that addressing those problems was outside of the
scope of this particular set of work.

Beyond the review comments, other m-of-n features are missing as
well.  These include:

 o Refactor TransactionSign.cpp so checks are as thorough in
   submit_multisign as they are in submit.

 o Add an RPC interface that returns the SignerList(s) associated
   with an account.

 o Potentially add a config-based flag to disable not-yet-ready
   features (in addition to the compile-time #if's).

 o Lay ground work for multiple SignerLists that support some kind
   of access control model.

 o Extensive testing.