get_signingaccount submit_multisigned errors [RIPD-480]:
Improved the error handling for these two RPC commands so they
return good error messages under the same circumstances that
the 'sign' and 'submit' RPC commands handle well.

The situation is improved.  But, to be honest, I'm not very pleased
with the current approach.  The error checking is distributed in
several functions with order dependencies between the calls.  It
makes the error handling difficult to visualize through code
inspection.

The expanded unit test helps.  It should help folks see consequences
when they change the error handling in the future.

If there were sufficient business payoff, I'd like to investigate
a data-driven field check approach.  That would allow the fields
being checked to be easily visualized.  An approach like that could
be especially valuable if it could apply to both RPC and transaction
handling.  But I don't believe I can justify that kind of change
right now.