Allow an account to be exclusively multisigned (RIPD-182):
The capabilities of multisigning are intended to be roughly
comparable to those of a regular key.  It is currently possible
to make an account that can only be signed using its regular key.
This is done by disabling the master key.  Since that's true, it
seems like it should be possible to do the same if an account is
multisignable (because it has a SignerList).

However, rippled goes to some effort to try and make sure an
account is always signable through some means.  So, for example,
rippled prevents a user from disabling their master key unless they
have a regular key in place.  Similarly, if an account's master key
is disabled rippled will not allow the user to remove their regular
key.  They can change the regular key, but they may not remove it.

To account for multisigning, and still provide comparable safety,
here are the rules this pull request employs:

 o An account can always add or replace a regular key or a
   SignerList as long as the fee and reserve can be met by the
   account.

 o The master key on an account can be disabled if either a
   regular key or a SignerList (or both) is present on the account.
   Either the regular key or the SignerList can be used to
   re-enable the master key later if that is desired.

 o The regular key on an account may only be removed if either the
   master key is enabled or the account has a SignerList (or both).

 o The SignerList on an account may only be removed if either the
   master key is enabled or a regular key is present (or both).