Add signer_list flag to account_info response (RIPD-1061):
Recovering the SignerList from an account may be a common problem
for API users.  This change is intended to help by adding a

"signer_list": true

flag to the account_info RPC command response if the account has
a signer list.  This allows the caller to determine whether they
need to collect the signer_list using the account_objects RPC
command.