[FOLD] makeAcctRoot returns a tl::expected<AcctRoot, NotTEC>:
The previous version of makeAcctRoot(), which returned a
std::pair<NotTEC, AcctRoot>, required constructing and returning an
unusable AcctRoot whenever there is a construction error.  The C++
standard proposal for std::expected...

http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2021/p0323r10.html

...looks like the perfect solution.

The proposal is not yet in the standard.  But the paper is live
and being reviewed by LWG, so it's looking promising for inclusion
in C++23.  The C++23 version looks very much like a slightly enhanced
std::optional.  The enhancement is that std::expected can tell the
caller why the call failed.

There is a boost::expected class, but it does not look very similar
to what is proposed for C++23.  So that one is not being used.

There is an open source header only implementation of std::expected
made available by Sy Brand.  I've copied that file and dropped it
into our source repository mostly unmodified.  This will give us
access to std::expected's capabilities prior to C++23.  Once we adopt
C++23 then we can change the #include statements and namespace names
and easily switch over the use the standard implementation.