Introduce tl::expected into the code base:
There's a C++ standard library proposal that is on track to be
available in C++23: std::expected.  The library looks like it
will be really useful in this code base.  The newly added file
is an open source implementation of the proposed standard
facility.  We can use this implementation, tl::expected, until
the new facility becomes available in C++23.  At that point
in time switching from the tl:: to the std:: implementation
should be pretty straight forward.